requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
matplotlib
shap
joblib
dotenv
google-generativeai